% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batch_script.r
\name{make_batch_script_template}
\alias{make_batch_script_template}
\title{Make a batch script to submit a SLURM job}
\usage{
make_batch_script_template(
  specname,
  slurm_jobname,
  walltime,
  partition,
  mem,
  ncores,
  array,
  offset,
  logname,
  nsims,
  nsteps,
  add_arrivals,
  rootdir,
  simdir,
  rscript_exec = "/users/jgantenb/build-sources/R-4.3.2/bin/Rscript",
  rscript_exec_args = "--vanilla",
  rscript_file,
  make_submit = TRUE,
  ...
)
}
\arguments{
\item{specname}{A character string giving the name of the job specification.
This parameter will be used to direct logs and output to the
correct locations.}

\item{slurm_jobname}{A character string giving the name of the job. Will be
appended to specname in the SLURM call.}

\item{walltime}{A character string giving the walltime for the job.}

\item{partition}{A character string giving the partition for the job.}

\item{mem}{A character string giving the memory for the job.}

\item{ncores}{An integer giving the number of cores for the job.}

\item{array}{A character string giving the array ID range.}

\item{offset}{A character string giving the offset that will map the array
IDs to the correct simulation number. Environment variable.}

\item{logname}{A character string giving the naming format of the log file.
Log files will be written to the .logs directory within the
directory specified by specname.}

\item{nsims}{An integer giving the number of simulations runs for the job.
Usually, this is set to 1 because we run 1 simulation per array
ID. Environment variable.}

\item{nsteps}{An integer giving the number of time steps in the simulation.
Environment variable.}

\item{add_arrivals}{A numeric input indicating the number of arrivals to add
per timestemp. Chosen in preliminary simulations to
maintain the population size at approximately 20,000.
Environment variable.}

\item{rootdir}{Full path to the project root directory.}

\item{simdir}{Full path to the directory in which to write the
simulation output. This should typically be a directory within
the ~/scratch directory.}

\item{rscript_exec}{Full path to the \code{Rscript} executable.}

\item{rscript_exec_args}{Arguments to pass to the \code{Rscript} executable.
Defaults to "--vanilla".}

\item{rscript_file}{Path to the R script to execute relative to project \code{rootdir}.
Will be passed as a full path to the SLURM scheduler.}

\item{make_submit}{Boolean. If \code{TRUE}, will run \code{make_submit_file()}.}

\item{...}{Pass arguments to \code{make_submit_file()}.}

\item{batchid}{An integer giving the batch id for the job.}
}
\value{
A batch script that contains specifications for a SLURM job array.
This specification is referred to as a template and is used as the
basis for the job submission file written by \code{make_submit_file()}.
}
\description{
This function writes a batch script that contains
specifications for a job array that will run the
requested number of simulations.
}
