% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_submit_file.r
\name{make_submit_file}
\alias{make_submit_file}
\title{Submit batch jobs based on batch specification}
\usage{
make_submit_file(specname, specdir, batches, offsets)
}
\arguments{
\item{specname}{A character string that should be the same as the
\code{specname} argument provided to \code{make_batch_script()}.}

\item{specdir}{A character string giving the full path to the destination
directory for the script.}

\item{batches}{A list containing vectors of integers, where the position
in the list corresponds to the batch ID.}

\item{offsets}{A numeric vector giving the offsets for each batch.}
}
\description{
This function writes a bash script that submits
batch jobs to the SLURM scheduler based
on the output of \code{make_batch_script()}.
}
